Move C++ only headers in TH to hpp
