+++ aten/src/THC/CMakeLists.txt
@@ -82,7 +82,7 @@ INSTALL(FILES
           THCScanUtils.cuh
           THCSortUtils.cuh
           THCAllocator.h
-          THCCachingAllocator.h
+          THCCachingAllocator.hpp
           THCCachingHostAllocator.h
           THCDeviceUtils.cuh
           THCDeviceTensor.cuh
