+++ torch/csrc/cuda/Module.cpp
@@ -7,7 +7,7 @@
 #include <sstream>
 #include <TH/TH.h>
 #include <ATen/ATen.h>
-#include <THC/THCCachingAllocator.h>
+#include <THC/THCCachingAllocator.hpp>
 #ifdef WITH_NCCL
 #include <nccl.h>
 #endif
